db/snapshotsync: move blocksnapshots up to db/snapshotsync/blocksnapshots
Sibling of freezeblocks rather than nested under it, matching how
freezeblocks itself sits directly under snapshotsync.
